Removed redundant code. Separate validators from helpers.
