+++ validators/index.js
@@ -1,10 +1,1 @@
-const helpers = require('../lib/helpers');
-
-module.exports = {
-    isArray: helpers.isArray,
-    isArrayOrSingle: helpers.isArrayOrSingle,
-    isString: helpers.isString,
-    isInteger: helpers.isInteger,
-    isBoolean: helpers.isBoolean,
-    isPath: helpers.isPath
-};
+module.exports = require('../lib/validators');
